astakos: Fix unknown variable in templates
Fix issue with the login template, where in the case of the master auth
provider, the 'provider' variable is undefined in the included
templates. This resulted in the 'primary_login_prompt' to not be shown.

Also, remove all references to 'master_auth_provider' in the auth
templates. They should be using only the 'provider' variable which is
set correctly from the 'login_base' template.