astakos: Map provider to user attributes
Extended the 'AuthProvider' class with a dictionary, mapping provider's
attributes to AstakosUser attributes. Also, this dictionary dictates
whether the user can change an attribute, or the value of this attribute
is automatically set by the provider.

Also, update profile forms by making attributes that are forced by the
provider to be read-only.

Finally, if all attributes are forced by the provider the
'ThirdPartyCreation' form is not required. In this case, the user is
 the AstakosUser is automatically created with the profile that is
 provided by the provider.