deploy: Add PITHCAT_UMASK setting to snf-image
PITHCAT_UMASK is needed if Archipelago does not run as root.
If the above setting is missing, the pithcat becomes really slow.

Up until now, snf-deploy used /etc/default/snf-image as a template with
default snf-image settings commented. These settings were never
updated and as a result they were different from the ones the package is
shipped with. This patch, after package installation, moves original
/etc/default/snf-image to snf-image.orig and then overwrites it with
the template that includes only the settings that deploy wants to
change.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>