plankton: Expose image's backend information
Make Plankton backend expose information about the backend entity that
is holding the data of the image, which in the case of Pithos is an
Archipelago mapfile. The goal of this commit is to make Cyclades unaware
of Archipelago mapfiles, and just use the backend information that is
exposed by the Plankton backend.

The information that is exposed is the following:
* 'backend_id': The ID of the entity that is backing up the image, which
   in case of Pithos contains the name of the Archipelago mapfile. The
   mapfile was until now exposed in the 'mapfile' field, which has
   basically been renamed to 'backend_id'.

* 'backend_uri': The URI of the entity that is backing up the image, as
  is expected by snf-image. In case of Pithos, this URI is of the form
  'pithosmap://$mapfile/$size'.

Also, the 'mapfile' column of the 'Image' database table, has been
renamed to 'backend_id'.

Finally, this commit moves the creation of snapshot's mapfile name
inside the 'register_snapshot()' method of the PlanktonBackend.

Signed-off-by: Christos Stavrakakis <stavr.chris@gmail.com>