cyclades: Handle instant snapshot registration
In case of Pithos, the Cyclades code is using the Plankton backend to
register a Pithos object for the snapshot. This object is created in the
"UNAVAILABLE" state and then Cyclades issue a Ganeti job to create the
snapshot. When the Ganeti job completes, the state of the Pithos object
is updated accordingly.

However, some implementations of the Plankton backend could support
instantly creating snapshots without the need to asynchronously create
the snapshot in Ganeti. For example, the Pithos backend that is used by
Plankton could be initialized as an Archipelago peer and instantly
create the snapshot.

This commit extends Cyclades to distinguish between synchronous and
asynchronous snapshots by interpreting the state of the snapshot and
issuing the Ganeti job only in the case of snapshots that are in the
"UNAVAILABLE" state.

Signed-off-by: Christos Stavrakakis <stavr.chris@gmail.com>