astakos: Implement Identity V3 authenticate call
Implement the POST /v3/auth/tokens API call that is the minimum API call
that is required to perform user authentication using the Openstack v3
Identity API. Currently, we support only the 'token' and 'password'
authentication methods. For the 'password' authentication method we are
using the UUID of the user and its token as username/password
credentials, exactly as we do for v2.

Signed-off-by: Christos Stavrakakis <stavr.chris@gmail.com>